Rename some "token" members to "keyword"
Summary:
I am making the names of the data structures self-consistent because we will be publishing this as an API.

For example, we are inconsistent about naming fields that represent keywords like "if" as "if_token" or "if_keyword".  I'm standardizing on "keyword".

Lots more diffs like this to come.

Reviewed By: dabek

Differential Revision: D3880723

fbshipit-source-id: f8c5e617e3eacc03a93fabf7cf4981126253dc6b