Make error node consistent with other nodes
Summary:
An "error" node in the tree used to contain a list of syntax nodes as its child. But this is unnecessary, because first, in practice so far this has only ever held a single token, and second, because we already have a node that contains a list of arbitrary items. If we want to mark a list of arbitrary items as erroneous, we can make a list and then make its parent an error.

(When we have better error recovery we likely will have error nodes that contain more complex structures than just a token.)

This reduces the number of "special" node types to just three: list, token and missing.

Reviewed By: jopamer

Differential Revision: D3888890

fbshipit-source-id: bed6eef9331f859216acda5925784e70e7a83edc