Remove warn_error_A from OSS build
Summary: IMO this is causing more trouble than good -- with all the different versions of OCaml floating around on different OSes, and with our fairly poor CI coverage of them, we tend to have some warning somewhere that fails the build for some people for no good reason. (4.03 on OS X in particular continues to cause me grief.) I don't think we get enough PRs to justify this.

Reviewed By: dabek

Differential Revision: D3884617

fbshipit-source-id: d963f1b001b979bcd6a995bf41074f9fb429697d