Replace use of unify in various ops with subtyping instead
Summary:
Operations such as arithmetic and array indexing are treated specially
in the Hack type checker. Where arguments are expected to be of particular primitive types,
sometimes unification is used rather than subtyping. Subtyping is to be preferred:

(1) It supports upper bounds on generic parameters, opaque types and enums.
For example, I might have a type `MyInt` defined to be `as int`. This would not
unify with `int` but is a subtype of `int`.

(2) It's the notion of compatibility used when calling user-defined functions, and
so should be used also for built-in functions. For example, I can define
```
function MyXor(int $a, int $b):int { return $a ^ $b; }
```
and then write `MyXor($x, $y)` even if `$x` and `$y` have type `MyInt` as described above.
The same should be true of writing `$x ^ $y` directly.

(3) As we move to a more constraint-based implementation of type inference, subtyping is
the primary notion.

Reviewed By: jopamer

Differential Revision: D3892781

fbshipit-source-id: 3aa28bb601d8ccfb879e3deb5f742125b6c700a2