Parse enums in decl mode
Summary:
Currently, we don't parse any top level enums when parsing in decl mode. This means that enums declared in decl files are never named, and can't be used in non-decl files.

It could be argued that enums aren't really top level entities, but I feel like this is a feature we should support. It will also make decl mode parsing of all files as a quick parsing step possible.

Reviewed By: jopamer

Differential Revision: D3877138

fbshipit-source-id: 5f102ffef2d460b123c34384515f00046913b895