Bugfix; arguments to rewriters flipped
Summary: Looking into the rewriter for parse trees from the full fidelity parser, I noticed that the arguments for the helper function in `rewrite_post` and `rewrite_pre` were in the wrong order. This bug wasn't surfaced yet, because nobody uses these functions (yet). (Also, corrected indent that was off.)

Reviewed By: ericlippert, andrewjkennedy

Differential Revision: D3886402

fbshipit-source-id: 6b32a0c3e4c45f5e5dc648af68abb10ef2cf6619