Include dependent files in Lazy_check fanout, if they are open in IDE
Summary:
If I have both A and B open, and a change in A generates errors in B, then we should report it as soon as possible. Before that, this error would only be reported when someone starts editing B itself, or during full check (after file save).

TODO: there is more that we need to capture here (I think - I don't have the repro case at the moment). Since we don't fan-out the phase2 declarations, we are not computing the whole set of to_recheck2 (see in [[ diffusion/FBS/browse/master/fbcode/hphp/hack/src/server/serverTypeCheck.ml$484,487 | serverTypeCheck.ml ]]), so when we "filter it down just to IDE files" it might be incomplete anyway, because some files were not inside in the first place.

I'll deal with it in separate diff.

Reviewed By: jopamer

Differential Revision: D3891404

fbshipit-source-id: 6fcbe5964783e863bd7841d182e942e127adb309