Only add the additional libraries required by enabled extensions to the targets that need it
Summary:
Currently the defines, includes, and libraries are being added to every target defined after the extension is configured. This causes problems with one of my other diffs which moves extension configuration to be the first thing that gets done.
This diff makes it so that the extra defines, includes and libraries are only added to the targets that actually need them.

Reviewed By: aorenste

Differential Revision: D3859164

fbshipit-source-id: faa4c3850d882b6364217b856efd28316cebfa82