Add a missing test exercising to_recheck2 set in ServerTypeCheck
Summary:
As mentioned in {D3891404}, I was surprised that when I erase the contents of this set, all of our tests still pass.

I though the purpose of this set is to capture the errors caused transitively through inheritance and typing usage (e.g Bar uses FooChild::f() which is inherited from Foo::f(), and we changed the method in Foo). But it looks like in cases like this one, there is a direct typing dependency between Foo and Bar (though that might be something that wasn't the case in the past).

After some very tiresome bisecting over real diffs in www, one case I found out where it is necessary, is when Foo didn't exist in the first place, and we are only adding it. This is the test case that I added.

The problem then was making this visible user error - the errors were correct, because FooChild ended up being redeclared due to "always redecl things that failed decl before" rule. Failing decl was less common in the past, when we had assume_php = true and less strict files - now I had to resort to HH_FIXME's to make it show.

I'll work on IDE/lazy check version of this rule in next diff.

Reviewed By: oulgen

Differential Revision: D3898080

fbshipit-source-id: dcdbacb91142edbde2c7aa6d1a405fe81f40b051