Fix indentation and add suggestions
Storeups and Loadups had an "else if" with wring indentation.
VptrFixVisit added call "use(r)" in the "useHint" and moved variable to the
end of the structure.
Replaced check if vout is empty to add vmodify() instead of the entire
block.
Fixed comments in macros where the fixVptr was removed.