Properly re-attaches the LightProcess signal action after exec.
Summary:
LightProcess registers a sig action handler with sigaction(), and therefore it must be restored through sigaction in the destructor of ShellExecContext. This patch implements that as well as
updating ShellExecContext to use sigaction when it registers the default signal handler in the contructor.

All platforms were affected by this bug to the effect of LightProcess::SigChldHandler() was being called with only one argument and trash in the other two.

Fixes test/slow/ext_process/lwp.php for AArch64.
Closes https://github.com/facebook/hhvm/pull/7314

Differential Revision: D3821760

Pulled By: aorenste

fbshipit-source-id: 8bcabaed213b69576705787037f9ef7a4baf5010