Replace patch Vptr and fix loadups/storeups
Summary:
patchVptr was a function called inside many lowers to handle the memory ref. This function was removed to be a new optimization phase "fixVptrsForPPC64", that works as the lower phase, replacing and adding new VASMs, and keep it separated from lower phase.
The memory reference of loadups and storeups need to be handled on PPC64, as on this architecture there no instruction to handle a reference with base, index and displacement.
Closes https://github.com/facebook/hhvm/pull/7337

Differential Revision: D3850918

Pulled By: aorenste

fbshipit-source-id: b33ffbf0b29b81a640b60ad437e1c16c22dd68b6