Avoid PHP code consistently when C++ exception pending
Summary:
When a C++ exception is pending, we avoided running PHP catch {} blocks,
but not PHP finally {} blocks. Let's fix this and be consistent in avoiding
execution of PHP code.

This brings us one step closer to unification of Catch and Fault entries and
also makes unwindPhp() with pending C++ exception behave more like unwindCpp().
Note that we can't hoist the check to the beginning of unwindPhp() since
pending C++ exceptions may be created during tearDownFrame().

Reviewed By: markw65

Differential Revision: D3905196

fbshipit-source-id: 965309fd83f74a5eb83a551ef38a5e1dbc4bb2ff