Proper fix for hhbbc type tests under clang
Summary:
It wasn't a compiler bug after all. The problem was
`auto foo = std::initializer_list<...>{...};`: foo is copy-constructed from the
temporary on the rhs, but copying an initializer_list doesn't copy the
underlying object storage or extend its lifetime. The compiler can free that
memory as soon as the temporary is destructed. We've just been getting lucky so
far because gcc and older clangs haven't done that (at least in a user-visible
way).

Reviewed By: binliu19

Differential Revision: D3911385

fbshipit-source-id: ad28ea8ddb21896acdd4964530fce12a47bdfd31