Add a new class to support TOC
Summary:
This is the first pull request of many to use a special feature for ppc:
TOC.

TOC (table of contents) is a memory region pointed by r2 that
contains immediate values to be loaded throughout the jit execution.
That is currently using hhvm's data-block as the memory region.

In this first pull request the code is not yet integrated as all our changes
in one PR would make it too big to be reviewed. For now, it just implemented
the singleton class that manages what is stored in TOC, and the "limmediate"
function that decides wether it is good or not to use TOC to load a specific
immediate value.
Closes https://github.com/facebook/hhvm/pull/7347

Differential Revision: D3872412

Pulled By: mxw

fbshipit-source-id: 5ecda591f2b2dd2793d4d6507ceaae3b25e3042d