Set MADV_DONTDUMP for worker's shared memory
Summary:
`madvise()` lets you advise the OS to avoid including the shared
memory in a core dump. Since the shared memory is very large, dumping this
memory leads to very large core dumps that take awhile to write, even at 100%
IO. When you have lots of worker processes, this can easily bring a machine to
its knees for hours.

Before the Windows support diffs, we used to fork the worker processes. This
meant we only had to set `madvise()` once, before the forks. Now, we have to
call it again when each worker starts up.

Reviewed By: dabek

Differential Revision: D3913891

fbshipit-source-id: 7e49ad12fb43a763d629392c950c1bcd63510f31