Rename "diagnostics" to "value" in diagnostics response
Summary: RPC framework requires that subscription body is inside the "value" field. We don't need to mention "diagnostics" anywhere, since the fact that those are diagnostics is encoded in ID - ID is equal to the original subscription request ID, and editor will know the expected type of response based on type of request.

Reviewed By: jamesjwu

Differential Revision: D3915323

fbshipit-source-id: f52e70a5ef09e4778e9a3b4000d23fefd161b101