Parse and ignore toplevel statements
Summary:
Right now,  the declaration parser ignores toplevel statements by ignoring each part token by token.

This is an issue if you write code like:

    <?hh // decl
   /* Assume foo is defined */
    foo(async ($x, $y) ==> { await $x; });

Which is valid in strict mode, but not decl mode.

To fix this, when a word appears as a toplevel statement, we parse it as a statement, and then ignore the result completely. We also throw away any parsing errors generated by parsing it as a statement, since the decl parser ignores these anyways(see the comment in the code).

Similarly, we ignore "use" statements. In a later diff we'll need to revisit the use statements since they matter for quick parsing but not for decl mode.

Reviewed By: jopamer

Differential Revision: D3879154

fbshipit-source-id: 4f12401a888f4ab41f31f2bddae479f44f440136