Retrieve new client FD earlier - during the client provider check loop
Summary:
Since moving our Watchman client to subscription mode, we can't give the user
typecheck errors for the real current status of what's on disk (only what
has been pushed through the Watchman pipe so far). Thus, we need to implement
a synchronous mode for when the user wants such information (for example,
invoking hh_client in the terminal).

The strategy for such a synchronous mode will be - touch a unique
file, then wait for that touch to be pushed down by Watchman. Thus, we will
be blocking until that push arrives. However, we are only blocking on watchman
because we want to satisfy that client's request. This, if the client
disconnects, we no longer need to block on Watchman and can move on.

This is the first step towards that - get the client FD earlier, so it can
be selected on simultaneously with the Watchman FD. If the client
disconnects, just move on.

Differential Revision: D3903172

fbshipit-source-id: 9071a57ee80c4651416be6603d4fee1f0f26c878