fix json_encode output for vec and keyset
Summary:
json_encode with vec and keyset had the following problems: P56650097

the issue was that in `VariableSerializer`, the comma separating each element in the array was written in `writeArrayKey`, but for `vec` and `keyset` we only call `writeArrayValue`. So,

* for `vec` and `keysets` (and vec-like php-arrays) we now write the comma in `writeArrayKey`.
* I tried removing writing the comma from `writeArrayValue`, even for `dicts` and hashmap-like php-arrays. But that is not feasible.

Most of the changes are for sending the `ArrayKind` information into `writeArrayKey` and `writeArrayValue`. An alternative design would be to add a `ArrayKind` field to `ArrayInfo` struct, but i thought its preferable to keep the memory smaller -- let me know if that is better.

Reviewed By: kodafb

Differential Revision: D3861979

fbshipit-source-id: 392bdf98b29ab3b2d6bf28e57f6f9ada3d1151d5