Fix various issues with require extends/implements
Summary:
The code attempted to check the requirements at the point that the
trait/interface was defined, but it did so without actually adding a
dependency between the trait and the requirement. This meant that the
requirements weren't always checked at the point of definition.

In addition, checking the requirements at the point of definition
rather than the point of use adds false dependencies, which can cause
autoload cycles.

The simplest fix is to just wait until the requirements are used
before checking them.

Reviewed By: ottoni

Differential Revision: D3908985

fbshipit-source-id: 18595b0ef98bcda57ae00b6de9df5fd63457e239