Separate vm mcgen and tc modules
Summary:
Move the various places we read from vmstate in mcgen into the request
handlers. The goal here is to have the mcgen and tc modules be as independent
of the request context as possible. Unfortunately we cannot currently assert
that the request heap is untouched as static arrays are first allocated as
counted arrays and then made static. This is particularly common in the
simplifier.

Reviewed By: swtaarrs

Differential Revision: D3888967

fbshipit-source-id: 61f0250186ed49e09352813a5448564dbbb14d9c