Add mechanism for TC branch profiling
Summary:
When enabled, the branch profiling adds to every vasm branch target a
sampling counter tick and a branch to a helper which logs information
about the edge that was taken.

Branches are identified by compile-time Func and Class, runtime Class
context, bytecode, IR opcode, IR offset, and vasm offset.

Reviewed By: kodafb, markw65

Differential Revision: D3732473

fbshipit-source-id: 1d041af63a97f26c8404c91f7135aa3a6efc00a0