Synchronous watchman query
Summary:
This is the last stage for correctness on moving towards watchman
subscriptions. When invoking hh_client in the terminal, we want to attempt
to get the current state of the file system. We accomplish this by touching a
random file and seeing that change be pushed down the watchman subscription.

Reviewed By: jopamer

Differential Revision: D3925266

fbshipit-source-id: 7757e7fe85685d6b83968ea7761c0b78154e067b