Fix PHP frame not saving the r31
The PHP frame was not saving the r31, which is a non-volatile register, and
it is used to save the SP value (like the r1).
This issue is visible only when the unwind runs, as it uses the r31 value
to find the previous frame. The old fix added an additional PHP frame to be
able to recover this value, but is was not the best one.
This diff assign a new register, the r26, to be the rvmfp, which makes the
r31 untouchable, so when unwinding, it will contains the correct value on
the last C++ frame.
Also in the inittc VASM, the r26 was initialized with the previous FP value,
which will be saved in the "stublogue{ true }", as enterTCHelper do not
create a new frame, and the r31 already has the same value of r1 (rsp),
otherwise the backchain would be in a loop.

Conflicts:
	hphp/runtime/vm/jit/vasm-ppc64.cpp