Add consolidated globalOptions type and parserOption and typecheckerOption views
Summary:
In order to take advantage of laziness, the parser heap needs to have access to parser options. But instead of threading parser options along with typechecker options
throughout the entire typechecker, it's easier and cleaner to consolidate the types and instead view them from different modules to limit their usage to their respective types.

This allows us to pass a single type t everywhere, but view the options as typecheckerOptions or parserOptions as necessary.

The diff isn't exactly what achow suggested, but I think it achieves the same goal. Tell me if using the module types approach would achieve something greater.

Reviewed By: alexchow

Differential Revision: D3919207

fbshipit-source-id: 6d565a3be20c5b3cbbacdf5691cef33cd496fa98