Do not silently ignore invalid file sync command from editor
Summary: yueli designed it initially to accept variety of malformed input. It makes debugging any issues impossible, since Hack's and editor's views of the world silently diverge. Eventually, we'll want to send a nice error and/or disconnect the non-conforming client, but for now we don't have this infrastructure. It's better to crash so we can develop the feature more easily.

Reviewed By: jopamer

Differential Revision: D3929480

fbshipit-source-id: 793b6ba30e99a1ce5d7a8ebd70c2efdae64b61fb