Sound checking of constraints for overridden generic methods
Summary:
The current scheme for checking that constraints on overriding methods (in the subclass)
are 'compatible' with the constraints on the overridden method (in the superclass)
is unsound. For example type holes see

See P56634518 (super constraints)
See P56644888 (as constraints)

The general principle for overridden method typing is the following: any
declared property of the overridding method must be a consequence of
declared properties of the overridden method, because a caller might be
calling through the superclass.

For constraints on generic methods, this means that if we have type parameters
in the superclass method

  <T1 cs1, ..., Tn csn> // Here cs1..csn are constraints

and

  <T1 ds1, ..., Tn dsn>

on the method in the subclass, then *assuming* constraints on the superclass,
constraints in the subclass must be a consequence. In the simplest case, if we
have

class C {
  public function foo<T as Derived>(...) ...
}

in the superclass then it's valid to write the 'weaker' constraint

class D extends C {
  public function foo<T as Base>(...) ...
}

in the subclass, because `T <: Derived` implies `T <: Base`.

Differential Revision: D3924359

fbshipit-source-id: 8f89d69db62ef7b601d6b6d3cc6df8f22c04e4c0