Add quick parsing step
Summary:
This diff adds a quick parsing flag to the parser, which allows it to parse and ignore function bodies. The parsed asts get added to the parsing heap in Decl mode, and then gets lazily updated to a Full parsing if the file is lazily typechecked.

For simplicity's sake, we still end up parsing everything, but just ignore the parsed output for function bodies. Thus, we don't save any initialization time, but writing a new toplevel parser for this might be the most straightforward way to do this.

Checking shared mem on my own devserver, we decrease the size of the shared heap after parsing from 4.9 GB to 2.1 GB.
Doing so saves a lot of memory on initialization; we get about a 2 GB memory save, and peak at 12 GB instead of 18.

https://fburl.com/443138467

This isn't complete yet as there's a few IDE problems, but I'll deal with that in a separate diff.

Reviewed By: oulgen

Differential Revision: D3859577

fbshipit-source-id: 6f45f0c135d54b5d48aaf16f36f7613693e70b48