Introduce emit{Set,Bind}L() helpers
Summary:
emit{Set,Bind}() requires the destination to be pushed on the stack before
the source, which could be inconvenient when the source is already on the
stack and the destination is a local.

Introduce a set of helpers for setting/binding the top of the stack to the
local. Update the existing workarounds to use these helpers. Fix a bug in
FinallyThunklet where an incorrect local was being unset.

Reviewed By: markw65

Differential Revision: D3927217

fbshipit-source-id: a34c91b7e3d18022e708aa0210f9f6cd3c7332d2