Kill hphp_context_init()
Summary:
It's just a wrapper for g_context.getNoCheck() (it probably did more
things in the past). I'm sad about the loss of symmetry between
hphp_context_{init,exit}() but I still think this is better. We might be able
to merge hphp_context_exit() and hphp_session_exit() with a little work; right
now there are very few places that call the former without calling the latter
immediately after.

Reviewed By: markw65

Differential Revision: D3941287

fbshipit-source-id: f10815cba33aeb7bd0cd06afc22714eef9e134d8