Speed up printir et al
Summary:
Running hhvm with TRACE=printir seemed really slow. perf report showed
that more than 90% of the time was spent in folly::Symbolizer. We
switched to folly::Symbolizer recently because it has a much smaller
footprint than the 5G+ used by bfd for doing the same thing, but
unfortunately, its very slow.

When we use the xed disassembler, it passes every address to
getNativeFunctionName, hoping to get a symbolic value back - but
nearly all the addresses it passes are in the tc, and will never have
a name.

This diff just excludes tc addresses from the search, and gets a 5x
speed up with TRACE=printir enabled.

Reviewed By: ricklavoie

Differential Revision: D3939328

fbshipit-source-id: 08036e22a1808d83e0d79a7aad4e3c0aff4b0e3f