Don't write-protect surprise flags in AssertVMUnused, rename to jit::VMProtect
Summary:
We still do request-local allocations in the jit which can set
surprise flags for OOMs (or trigger the GC, which also sets surprise
flags). Ignore Eval.MapTgtCacheHuge in !NDEBUG builds because that would change
the page size.

While I was in here I also gave it a more sensible name, moved it to jit/, and
cleared up the rules around nesting.

Reviewed By: markw65

Differential Revision: D3940056

fbshipit-source-id: 5fe5dc381f887e709200d3d6a96dd140a96dddfe