Convert one more Not_found into Decl_heap_elems_bug
Summary: {D3862080} was supposed to convert this random crash into a specific exit code, but the logs show that it's still happening, and in the last release there are NO instances of Decl_heap_elems_bug. The reason: one more find_unsafe that I missed. Not sure if that means that there is some logic wrong with constructors in particular, or just are constructors always checked first and this si why they show up.

Reviewed By: oulgen

Differential Revision: D3946796

fbshipit-source-id: 5f2e0f5b1c307a4afa2918d55c444d34e8f437a9