Helper function to find classes using elements being invalidated
Summary:
Lazy decl is not lazy enough - it assumes that class is declared, or it's not along with ALL of it's descendants. If a situation where E extends D extends C arises, and C and E are not declared, lazily declaring E will not re-declare the C elements that are shared with E, so the result will be incomplete. If we have such hierarchy, and we want to remove C, we need to remove all of it's descendants too. This adds a utility function that can find such classes.

See D3941897 for usage.

Reviewed By: alexchow

Differential Revision: D3941887

fbshipit-source-id: 95be36a3fcce806019383bc01adb3f29716cc723