Add field prefixes to schema
Summary:
OCaml requires that record field names be unique in a module; we therefore prefix every field name with a string that makes it unique.

This is going to be irritating when I generate APIs for Javascript and Hack, so I've added the prefix to the schema so that we can apply it programmatically, or not, as we see fit.

This necessitated a few small changes in some of the field names.  I've made notes in the text of places where field names are inconsistent and I'll fix those up in a later diff.

Reviewed By: dabek

Differential Revision: D3951517

fbshipit-source-id: 9d2b15ba03ea359dcebcb7af6a897b1a4199e977