Add JobQueue option for conservative worker growth, to be used for admin server
Summary:
The admin server is currently single-threaded by default, which is usually fine as its traffic is expected to be low, but could become a problem with long-running requests (or hung requests due to bugs).

One reason for not raising its max thread count, is it would usually be a waste of resources, when the second thread might be spwaned just by chance of two requests arriving close together.

Add therefore an option to JobQueue to make its thread growth more conservative: rather than the default of targeting one new worker for every one queued job,
scale the queue length by a given ratio. (If there are no workers, the division is rounded up to target at least one worker for a non-empty queue.)

The max thread count cap is still applied as before.

Reviewed By: markw65

Differential Revision: D3950960

fbshipit-source-id: 05aed6d53b4c963a4c2b5bfac1b69d7d214d05b2