Do not marshal ocaml strings
Summary:
After {D3941748} how objects are serialized to the shared heap is now the responsibility of `hh_shared.c`. One thing that is obvious is that if we have an Ocaml string, there is no need to marshal it. Instead for can write the string directly into the shared heap. This will save space in the heap because we avoid the 20+ byte header that Ocaml writes out when marshaling a value.

As part of this I needed to steal some bits from the `size_t` we store as a header for our heap allocated objects. I limit it to 31 bits and use an additional bit to indicate if it is a raw string or a serialized object. I leave 32 bits as aux data that I will use in a following diff.

This optimization will be useful when we intern strings in the heap.

Depends on D3941748

Reviewed By: jamesjwu

Differential Revision: D3941839

fbshipit-source-id: c622df64264e6289ec3269dfb1bb86c4bb5a0303