Only use provided class name with __PHP_Incomplete_Class in debug / serialize contexts
Summary:
This is a partial revert of D821936: I'm changing the conditions in which we trust the class name to be more restrictive. I'm making the change for a few reasons:

1. It is incompatible with upstream PHP behavior (https://3v4l.org/CVUKh).
2. It is incredibly confusing because it can hide the fact that you have an instance of `__PHP_Incomplete_Class_Name` when trying to debug objects using `var_dump` or `print_r`.

I am deliberately preserving the behavior on serialize calls because if someone unserializes a class we don't know about and then serializes it again they should get back what they put in.

Reviewed By: markw65

Differential Revision: D3869483

fbshipit-source-id: 9016f74d72ed58e6724a9e741493e3c27dfdafa1