Don't register rds bases created for VMProtect
Summary:
We iterate the registered TLS bases and read profile data from their local
sections. Doing this to the pseudo-bases we create for VMProtect causes us to
fault if they were unregistered by the time we got to them (we don't hold the TLS
base lock while iterating the list, we just grab it briefly in rds.cpp when we copy
the list).

Reviewed By: swtaarrs

Differential Revision: D3952292

fbshipit-source-id: a543941b2f5969b35a403908c61e5060572eb9f7