Compress contents of heap
Summary:
Compressing objects before copying in the heap reduces the size of our heap after intializing by ~40%. lz4 is also a fast enough compression algorithm that it is neglible impact on performance. To do this I use the remaining 32 bits in `hh_header_t` to indicate the expected size of the data after it is uncompressed. If the size is 0, then the data is assumed to be uncompressed already.

This requries adding a dependency on lz4. HHVM already has a dependency on lz4 so I updated the Hack OSS build to require building lz4 in order to build Hack.

Reviewed By: jamesjwu

Differential Revision: D3941923

fbshipit-source-id: 08031606b8726864dc0a4e12fe606d6306c43d76