Introduce local cache for full parsing heap, don't store full trees in shared memory
Summary:
In order to fully realize the benefits of quick parsing, we want to *never* store full ASTs in sharedMem, and instead lazily load it as necessary.

This diff starts at this process by introducing a thread level hashtable to store full ASTs, to avoid storing ASTs in the regular ParserHeap.

We keep the "Full" and "Decl" modes of the ParserHeap because many users of the ParserHeap(build, the IDE, etc) have the logic where they store a full AST into the ParserHeap and then take it back directly.

In a future diff we can make all of these lazy as well, but this diff gets most of the memory gains without rewriting all of the parser build logic.

This graph shows a memory savings of running a full, non-ministate typecheck on www with lazy parsing off and on. Assuming we end up parsing all of www, we still save about 2.5 GB of shared memory.
https://fburl.com/448941502

Just loading from saved state lazily, with this diff, lazy parsing gets the sharedHeap size down to ~ 4GB.

Reviewed By: dlreeves

Differential Revision: D3932439

fbshipit-source-id: cd833d8f1e249e32baedb085af4fe37dc7a372a4