Making rewriter return optional to allow surjective rewrites
Summary: This diff is preparatory; the rewrite from the new full fidelity parse tree to the old pre-nast is surjective (if only for the discarding of all the trivia, i.e. whitespace). The rewriter only allowed for arity-preserving rewrites. This diff allows rewriters to drop branches by rewriting them to `None`.

Reviewed By: ericlippert

Differential Revision: D3887050

fbshipit-source-id: 2ee42140de59e1cf762c01532ed35757d8eea1cd