[AArch64] Add map-flags pass.
This new pass tracks every def of a VregSF and the uses. Then the
pass iterates over each use of a VregSF and updates the new field
in the instruction that defs the VregSF with the individual,
platform-specific status flags which need to be set.

The pass aims to accomplish the following:
  A) Allows platforms to map the condition codes at a use instruction
     based on the def instruction. This is especially beneficial for
     platforms that don't have a perfect 1:1 mapping of condition
     codes in terms of semantics.
  B) Allows platforms to optimize their emitters to only set the
     status flags bits that are necessry. This avoids emitting
     expesive code to synthetically set status flags unless
     absolutely required.

Fixes:
  test/slow/php7_bcbreak/7/int_semantics/nan_inf_cast.php