Fix GP args to include r10 on PPC64
Summary:
Somehow only r3-r9 were defined although the ABI points out that r10 is
also a GP arg. This didn't bring problems to PPC64 before as it's rare
to call a function with so many arguments however the msg_receive
function is one of them and the garbage it used was valid for most of
the time.  (The issue happened when installing hhvm to
/usr/local/bin/hhvm and running hphp/test/run)

When calling a function with 8 GP args, HHVM was spilling the 8th
argument onto the stack (and keeping 16-bytes alignment, of course)
before setting all the other register arguments:
addi    r3,r31,-128
addi    r27,r27,-8
stdu    r3,-8(r27)

But msg_receive was reading from r10 even if it was undefined, therefore
garbage was being read. Now it correctly sets it as:
addi    r10,r31,-128
Closes https://github.com/facebook/hhvm/pull/7395

Differential Revision: D3944683

Pulled By: mxw

fbshipit-source-id: d69370122fcd9bb22dcdc678fc0057df42dba141