Improve decompose_subtype to iterate over transitive constraints in environment
Summary:
For safe `instanceof` we take an assumed constraint

  ty1 <: ty2

and add to the environment bounds on generic type parameters that must hold for ty1 <: ty2 to be true.
For example, if the assumed constraint is

  C<T1> <: C<T2>

for a covariant class C, then we add

  T1 <: T2

to the environment.

Sometimes these bounds arise transitively, and we currently miss them. For example, the environment may
already contain the lower bound

  C<T1> <: T

and then we have an assumed constraint

  T <: C<T2>

so we add this as an upper bound on T.

But we should also add

  T1 <: T2

to the environment, as this is a consequence of the two bounds.

Note that although this currently only affects obscure uses of `instanceof` (see test), we are now
in a position to use `SubType.add_constraint` whenever we need to assume a general constraint between
two types. This will be the subject of a future diff.

Reviewed By: jopamer

Differential Revision: D3960666

fbshipit-source-id: d30e3feb791e21fcb22f431f2c24860391518041