Match Units of closure Class* and containing Func* in CreateCl
Summary:
If multiple Units are created for the same file contents (which can
happen, e.g., if we swap out the file for a different version, and then
swap it back), then we can end up with multiple different Classes in
the clsList() for the closure.  This could lead to weird races---e.g.,
if the two versions of the Unit are U1 and U2, while compiling the
CreateCl for U1, U2 might have finished merging, and we might end up
with the Class* for U2 instead of the Class* for U1.

Fix this by walking the clsList() until we find a Class* whose Unit
matches our own.

Reviewed By: markw65

Differential Revision: D3963875

fbshipit-source-id: 7fbacda025dd8cba7d68462384c2602110b90fa0