Generalize where constraints to pairs of types
Summary:
The syntax of `where` constraints supports general types on either side of
the `as`, `super` or `=` symbol, but currently it is enforced that the
left-hand type is a generic parameter. But instantiation through the
superclass (e.g. `class D extends C<string>`) means that we have to
consider general constraints, in particular to implement sound rules
for method overriding. It turns out to be simpler to support this at
the user level too, so we now permit syntax such as

  function foo<Tu>() where Cov<T> as Cov<Tu>

and the Hack type checker will decompose the constraints appropriately.

Reviewed By: jopamer

Differential Revision: D3961012

fbshipit-source-id: a1598ff4f1dc528bad188758df6a2f3d8130c69f