Explicitly disable type-scanners on Clang
Summary:
Previously there was no need to explicitly disable type-scanner
generation on Clang, because one could just not define HHVM_BUILD_TYPE_SCANNERS
on it. However, now that we're going to move to a model where it will be on by
default, it makes sense to start explicitly excluding it.

Reviewed By: mxw

Differential Revision: D3973061

fbshipit-source-id: fb4eb1afec90717f97c1948d5618e9425de79649