Fix AliasedNamespaces
Summary:
The settings specified in the various config files included a leading
backslash, which is never correct, and failed various assertions.

I'm surprised the code even worked, because there are a number of code
paths where we don't normalize the name (because we know the parser
did it for us); but apparently there are at least some paths where it
does work.

This normalizes the names (and also drops invalid ones) at startup
time.

Reviewed By: paulbiss

Differential Revision: D3970107

fbshipit-source-id: 31fe2ba2810f11e8ae48e3a30bdff3c77d09b274