Always conservative type-scan folly::Optional
Summary:
The representation of folly::Optional recently changed, and now it
contains an union with multiple types. As a result, the type-scanner can no
longer automatically generate a type-scanner for it. Normally we would fix this
with annotations, but folly::Optional is not part of our code base, thus we
cannot add any. As a stop-gap measure, hard code folly::Optional to always be
conservative scanned. While we're at it, move std::shared_ptr there as we have
similar issues with that.

Reviewed By: mxw

Differential Revision: D3977747

fbshipit-source-id: e2c09cd88f49eae31b62ce6b4a56e63d485037d4