Invalidate class element users when invalidating a class
Summary: As described in previous diff, D3941887, when removing class C we need to invalidate all of it's descendants too. This makes the lazy incremental mode much less efficient than I planned to, and we'll eventually want to have a smarter lazy decl that doesn't require doing it. For now adding this code since I want to have a correct working prototype before focusing on fine tuning performance in degenerate cases.

Reviewed By: alexchow

Differential Revision: D3941897

fbshipit-source-id: a11c1aa6e8c4dc41ae86ba71469787af5fc5f851