Fix use after free in async generators
Summary:
An async generator can finish with the last reference owned by the asio
scheduler. Processing of return event on an async generator removes a circular
reference between the async generator and its associated wait handle.

This diff fixes a bug when AsyncGenerator::m_waitHandle field was used to
decref the wait handle right after the AsyncGenerator object was freed. We now
use a local to hold onto the wait handle and use req::ptr<> to properly
maintain ownership.

Reviewed By: mxw

Differential Revision: D3975961

fbshipit-source-id: dbb8676b43156a338c27f51ab3080a5405062199