Arithmetic operations and subtyping
Summary:
There are a number of bugs associated with arithmetic operators (+, -, *, /, etc.)
The current implementation inspects inferred types for operands directly in order
to dispatch on type for overloading. Better is to use subtype tests, as subtyping
deals with "as" constraints (on generics, on enums, etc.), and also deals with
unions with no need for special casing in the code.

There was a bug in `Typing_subtype.is_sub_type` caused by `HH_FIXME` declarations.
See comment in code for explanation of the fix.

The rules governing "any" propagation are intricate and somewhat arbitrary. I've tried to preserve existing behaviour, and have a test file (arithmetic_any.php) that exercises most of the combinations.

Reviewed By: jopamer

Differential Revision: D3968065

fbshipit-source-id: 4a80769d4eb6fbeabdb7037397a46b3e9e5e13a0