Add HeaderKind::Closure & ClosureHdr, remove builtinODTailSize & builtinObjSize
Summary:
Most objects are laid out like this:

  [ObjectData][property 0][property 1]...

However, closure instances are laid out like this:

  [ObjectData][m_ctx][property 0][property 1]...

builtinODTailSize gave us the offset from the end of ObjectData
to the first property, and was the same value as builtinObjSize.
Since most builtins have been converted to native-data, only
Closure needed special treatment. so:

1. Identify closure instances with a new HeaderKind
2. Add a new header for closures to store m_ctx before the closure's
ObjectData, like other NativeData objects. (even though closure isn't NativeData).
3. Also use ClosureHdr to store the full size of the closure, so we don't need
to recompute it on free.
4. eliminate builtinODTailSize & builtinObjSize, since now they're always 0.

The new layout for closures is like this:

  [ClosureHdr][ObjectData][property 0][property 1]...

Mostly this resulted in mechanical changes everywhere, but it exposed a couple
bugs in ext_heapgraph, where it did the right thing for plain Objects, but not
special-layout objects (NativeData, waithandles, collections, and now closures).

Reviewed By: markw65

Differential Revision: D3975629

fbshipit-source-id: e9e3eb81d48416dec3069bd63d29ddd97a43bfa1